+++ src/main/kotlin/models/IMotorGasolina.kt
@@ -1,8 +1,12 @@
 package dev.carloscy.models
 
+/**
+ * Interfaz que representa el motor de un vehículo de gasolina.
+ * @property consumo consumo de combustible del motor.
+ */
 interface IMotorGasolina: IMotor {
 
-    var consumo: Double
+    var consumo: Int
 
     override fun arrancar() {
         println("Arrancando motor de gasolina.")
